feat: Update Astro site with dynamic blog post list
This commit updates the Astro site by dynamically generating the list of blog posts on the blog page. It replaces the hardcoded list with a loop that iterates over all the blog posts and displays them with their respective titles and URLs. This change improves the maintainability of the site and allows for easier addition and removal of blog posts in the future.